+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -15,6 +15,7 @@ public abstract class Participant {
 
     abstract boolean canHit();
     public abstract List<Card> showStartCards();
+
     public abstract String getName();
 
     public boolean isBust() {
