+++ src/test/java/blackjack/domain/BlackjackGameTest.java
@@ -52,4 +52,63 @@ class BlackjackGameTest {
         assertThat(score).isGreaterThan(16);
     }
 
+    @DisplayName("플레이어들의 수익을 각각 계산한다.")
+    @Test
+    void testDealerProfit_playersProfit() {
+        // given
+        CardHand playerHand= new CardHand();
+        playerHand.add(new Card(CardSuit.CLUB, CardRank.KING));
+        playerHand.add(new Card(CardSuit.CLUB, CardRank.JACK)); // 20
+
+        int player1Bet = 10000;
+        Player player1 = TestUtil.createPlayerOf(playerHand, player1Bet);
+
+        int player2Bet = 5000;
+        Player player2 = TestUtil.createPlayerOf(playerHand, player2Bet);
+
+        CardHand dealerHand = new CardHand();
+        dealerHand.add(new Card(CardSuit.CLUB, CardRank.KING));
+        dealerHand.add(new Card(CardSuit.CLUB, CardRank.ACE)); // 21
+        Dealer dealer = new Dealer(dealerHand);
+
+        List<Player> players = List.of(player1, player2);
+
+        // when
+        BlackjackGame game = new BlackjackGame(players, dealer, CardDump.shuffledDump());
+        Map<Player, Integer> playersProfit = game.calculatePlayersProfit();
+
+        // then
+        assertThat(playersProfit.get(player1)).isEqualTo(-10000);
+        assertThat(playersProfit.get(player2)).isEqualTo(-5000);
+    }
+
+    @DisplayName("딜러의 수익은 플레이어수익 합의 반대이다")
+    @Test
+    void testDealerProfit_playersProfitSumReverse() {
+        // given
+        CardHand playerHand= new CardHand();
+        playerHand.add(new Card(CardSuit.CLUB, CardRank.KING));
+        playerHand.add(new Card(CardSuit.CLUB, CardRank.JACK)); // 20
+
+        int player1Bet = 10000;
+        Player player1 = TestUtil.createPlayerOf(playerHand, player1Bet);
+
+        int player2Bet = 5000;
+        Player player2 = TestUtil.createPlayerOf(playerHand, player2Bet);
+
+        CardHand dealerHand = new CardHand();
+        dealerHand.add(new Card(CardSuit.CLUB, CardRank.KING));
+        dealerHand.add(new Card(CardSuit.CLUB, CardRank.ACE)); // 21
+        Dealer dealer = new Dealer(dealerHand);
+
+        List<Player> players = List.of(player1, player2);
+
+        // when
+        BlackjackGame game = new BlackjackGame(players, dealer, CardDump.shuffledDump());
+        int dealerProfit = game.calculateDealerProfit();
+
+        // then
+        assertThat(dealerProfit).isEqualTo(15000);
+    }
+
 }
